+++ IntegrationBus/source/mw/fastrtps/idl/FlexRayTopicsPubSubTypes.cxx
@@ -1277,6 +1277,109 @@ namespace ib
                     return true;
                 }
 
+                CycleStartPubSubType::CycleStartPubSubType()
+                {
+                    setName("ib::sim::fr::idl::CycleStart");
+                    m_typeSize = static_cast<uint32_t>(CycleStart::getMaxCdrSerializedSize()) + 4 /*encapsulation*/;
+                    m_isGetKeyDefined = CycleStart::isKeyDefined();
+                    size_t keyLength = CycleStart::getKeyMaxCdrSerializedSize()>16 ? CycleStart::getKeyMaxCdrSerializedSize() : 16;
+                    m_keyBuffer = reinterpret_cast<unsigned char*>(malloc(keyLength));
+                    memset(m_keyBuffer, 0, keyLength);
+                }
+
+                CycleStartPubSubType::~CycleStartPubSubType()
+                {
+                    if(m_keyBuffer!=nullptr)
+                        free(m_keyBuffer);
+                }
+
+                bool CycleStartPubSubType::serialize(void *data, SerializedPayload_t *payload)
+                {
+                    CycleStart *p_type = static_cast<CycleStart*>(data);
+                    eprosima::fastcdr::FastBuffer fastbuffer(reinterpret_cast<char*>(payload->data), payload->max_size); // Object that manages the raw buffer.
+                    eprosima::fastcdr::Cdr ser(fastbuffer, eprosima::fastcdr::Cdr::DEFAULT_ENDIAN,
+                            eprosima::fastcdr::Cdr::DDS_CDR); // Object that serializes the data.
+                    payload->encapsulation = ser.endianness() == eprosima::fastcdr::Cdr::BIG_ENDIANNESS ? CDR_BE : CDR_LE;
+                    // Serialize encapsulation
+                    ser.serialize_encapsulation();
+
+                    try
+                    {
+                        p_type->serialize(ser); // Serialize the object:
+                    }
+                    catch(eprosima::fastcdr::exception::NotEnoughMemoryException& /*exception*/)
+                    {
+                        return false;
+                    }
+
+                    payload->length = static_cast<uint32_t>(ser.getSerializedDataLength()); //Get the serialized length
+                    return true;
+                }
+
+                bool CycleStartPubSubType::deserialize(SerializedPayload_t* payload, void* data)
+                {
+                    CycleStart* p_type = static_cast<CycleStart*>(data); //Convert DATA to pointer of your type
+                    eprosima::fastcdr::FastBuffer fastbuffer(reinterpret_cast<char*>(payload->data), payload->length); // Object that manages the raw buffer.
+                    eprosima::fastcdr::Cdr deser(fastbuffer, eprosima::fastcdr::Cdr::DEFAULT_ENDIAN,
+                            eprosima::fastcdr::Cdr::DDS_CDR); // Object that deserializes the data.
+                    // Deserialize encapsulation.
+                    deser.read_encapsulation();
+                    payload->encapsulation = deser.endianness() == eprosima::fastcdr::Cdr::BIG_ENDIANNESS ? CDR_BE : CDR_LE;
+
+                    try
+                    {
+                        p_type->deserialize(deser); //Deserialize the object:
+                    }
+                    catch(eprosima::fastcdr::exception::NotEnoughMemoryException& /*exception*/)
+                    {
+                        return false;
+                    }
+
+                    return true;
+                }
+
+                std::function<uint32_t()> CycleStartPubSubType::getSerializedSizeProvider(void* data)
+                {
+                    return [data]() -> uint32_t
+                    {
+                        return static_cast<uint32_t>(type::getCdrSerializedSize(*static_cast<CycleStart*>(data))) + 4 /*encapsulation*/;
+                    };
+                }
+
+                void* CycleStartPubSubType::createData()
+                {
+                    return reinterpret_cast<void*>(new CycleStart());
+                }
+
+                void CycleStartPubSubType::deleteData(void* data)
+                {
+                    delete(reinterpret_cast<CycleStart*>(data));
+                }
+
+                bool CycleStartPubSubType::getKey(void *data, InstanceHandle_t* handle, bool force_md5)
+                {
+                    if(!m_isGetKeyDefined)
+                        return false;
+                    CycleStart* p_type = static_cast<CycleStart*>(data);
+                    eprosima::fastcdr::FastBuffer fastbuffer(reinterpret_cast<char*>(m_keyBuffer),CycleStart::getKeyMaxCdrSerializedSize());     // Object that manages the raw buffer.
+                    eprosima::fastcdr::Cdr ser(fastbuffer, eprosima::fastcdr::Cdr::BIG_ENDIANNESS);     // Object that serializes the data.
+                    p_type->serializeKey(ser);
+                    if(force_md5 || CycleStart::getKeyMaxCdrSerializedSize()>16)    {
+                        m_md5.init();
+                        m_md5.update(m_keyBuffer, static_cast<unsigned int>(ser.getSerializedDataLength()));
+                        m_md5.finalize();
+                        for(uint8_t i = 0;i<16;++i)        {
+                            handle->value[i] = m_md5.digest[i];
+                        }
+                    }
+                    else    {
+                        for(uint8_t i = 0;i<16;++i)        {
+                            handle->value[i] = m_keyBuffer[i];
+                        }
+                    }
+                    return true;
+                }
+
 
                 ControllerStatusPubSubType::ControllerStatusPubSubType()
                 {