+++ IntegrationBus/source/sim/fr/FrDatatypeUtils.hpp
@@ -20,6 +20,7 @@ bool operator==(const TxBufferConfig& lhs, const TxBufferConfig& rhs);
 bool operator==(const ControllerConfig& lhs, const ControllerConfig& rhs);
 bool operator==(const HostCommand& lhs, const HostCommand& rhs);
 bool operator==(const ControllerStatus& lhs, const ControllerStatus& rhs);
+bool operator==(const CycleStart& lhs, const CycleStart& rhs);
 
 std::ostream& operator<<(std::ostream& out, const FrSymbol& symbol);